tách export filter riêng
