+++ package.json
@@ -67,7 +67,10 @@
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-sass": "^1.2.2",
     "rollup-plugin-typescript2": "^0.29.0",
+    "sass": "^1.32.8",
+    "sass-loader": "^11.0.1",
     "styled-components": "^5.2.1",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.2",
     "ts-jest": "^26.5.0",
