+++ src/styles/filter.css
@@ -1,0 +1,45 @@
+.expand-filter-mobile{
+  transition: max-height 1s;
+  max-height: 1000px;
+}
+
+.shrink-filter-mobile{
+  max-height: 0;
+  transition: max-height 0.3s;
+}
+
+@keyframes rotateFilterExpand {
+  from{transform: rotate(0);}
+  to{transform: rotate(90deg);}
+}
+
+@keyframes rotateFilterShrink {
+  from{transform: rotate(90deg);}
+  to{transform: rotate(0);}
+}
+
+.accordion-filter-mobile-icon-expand{
+  animation: rotateFilterExpand 0.3s forwards;
+}
+
+.accordion-filter-mobile-icon-shrink{
+  animation: rotateFilterShrink 0.3s forwards;
+}
+
+@keyframes overflowVisibleFilter {
+  0% {overflow: hidden;}
+  100%{overflow: visible;}
+}
+
+@keyframes overflowHiddenFilter {
+  0% {overflow: visible;}
+  100%{overflow: hidden;}
+}
+
+.overflow-visible-filter{
+  animation: overflowVisibleFilter 0.5s forwards;
+}
+
+.overflow-hidden-filter{
+  animation: overflowHiddenFilter 0.5s forwards;
+}
